Reduce Archive Log retention for 2 days in post-snapshot hook (#14)
* Decrease Archive Log retention from 7 to 2 days to avoid VDB's halting due to exhausted recovery area.
This is judged to be quite safe, because it's a post-snapshot hook calling the action.
The current snapshot will contain all the new logs, even beyond 2 days, up to the point of exhuasting the FRA.
Only once a new snapshot completes will the logs get purged.  The new snapshot will still have all the old logs,
since it comletes prior to the purge.
This could probably be dropped further to 1 day.