Reduce Archive logs to 1 day (#15)
* Changing Archive Retention to 1 day
Reduce to keeping 1 day. Since the purge only runs as a post-snapshot hook, we assume the snapshots will still have the logs they need. The main constraint we're trying to solve is not overflowing the Flashback Recovery Area, which has a set configured size and halts the database when it's full, even though there is ample disk space available.